+++ src/cpu/idt.c
@@ -5,7 +5,7 @@
 idt_gate_t idt[IDT_ENTRIES];
 idt_register_t idt_reg;
 
-void set_idt_gate(int n, uint32 handler) {
+void set_idt_gate(int n, uintptr handler) {
     idt[n].low_offset = low_16(handler);
     idt[n].sel = KERNEL_CS;
     idt[n].always0 = 0;
@@ -14,7 +14,7 @@ void set_idt_gate(int n, uint32 handler) {
 }
 
 // Special function for syscall gates that need to be accessible from user mode (DPL=3)
-void set_syscall_gate(int n, uint32 handler) {
+void set_syscall_gate(int n, uintptr handler) {
     idt[n].low_offset = low_16(handler);
     idt[n].sel = KERNEL_CS;
     idt[n].always0 = 0;
@@ -25,7 +25,7 @@ void set_syscall_gate(int n, uint32 handler) {
 }
 
 void set_idt() {
-    idt_reg.base = (uint32) &idt;
+    idt_reg.base = (uint32)(uintptr)&idt;
     idt_reg.limit = IDT_ENTRIES * sizeof(idt_gate_t) - 1;
     __asm__ __volatile__("lidtl (%0)" : : "r" (&idt_reg));
 }
